Further split up manager.rs into submodules
